Add rule Scope configuration option to rules-based sampler (#440)
We'd like to be able to apply all of the conditions in a rules-based sampler rule to a single span. In other words, we'd like a rule to be considered matched only if all of the conditions in that rule are satisfied by a single span in the trace. This allows us to more accurately make per-service sampling decisions in a dataset written to by multiple services.

Co-authored-by: Mike Goldsmith <goldsmith.mike@gmail.com>