maint: change title on release workflow from (unset) var to hard-coded string (#474)
Error in previous workflow that attempted to create an issue states "Input required and not supplied: title".
While the title does match that of the example provided in the actions repo, it is based on a variable that is set in the example that we do not have set. I believe this is causing a blank title which is not allowed, as the title value is required. This PR changes the title from a variable that is unset ({ steps.date.outputs.today }}) to a hard-coded string (Bump Refinery to Latest Version).