Fix variable shadowing bug (#519)
## Which problem is this PR solving?

The tip of refinery was crashing in metrics with a nil pointer error. It was puzzling why a function that was supposed to never return nil was in fact returning nil. The problem was a subtle variable shadowing issue that I only found by writing a test script and running it in the debugger:

```go
	metric, ok = metrics[name]
	if !ok {
		// create new metric using create function and add to map
		metric := createMetric(name)
		metrics[name] = metric
	}
	lock.Unlock()
	return metric
```
The := on the createMetric created a new variable called `metric` that was scoped to the body of the if clause. The next line properly assigned it to the map and then threw it away and returned the nil pointer created on the top line.


## Short description of the changes

- Removed an unneeded `:`
- Added some tests for the getOrAdd function
- Fixed a couple of spelling errors and a linter nit