Cloud support (#521)
## Which problem is this PR solving?
When hosting Refinery in a cloud environment, certain network
assumptions are no longer true.
Issues arise from these assumptions while hosting in our cloud
environment using [Nomad](https://www.nomadproject.io/) as container
orchestration.

Terms, as used in Nomad:
- Task: Single deployed container.
- Group: Group of tasks running as a scalable service (e.g. Refinery).

This assumption is not necessarily true under Nomad:
- The IP assigned to a task's `eth0` can be accessed by other tasks in
the group.

Redis peering expects peers to talk to each other using their `eth0` IP.
This behavior can be overridden by specifying `RedisIdentifier` in
config. This value will be advertised as the peer's IP in the peer list.

However, on startup, peering uses its `eth0` IP to cross-reference in
the peers list found in Redis. Since the internal `eth0` IP is not the
true public IP (set by `RedisIdentifier`), this check will always fail
and the container exits in error.

## Short description of the changes
A workaround is to treat `RedisIdentifier` as the true public IP address
of the task. Thus, it should be one of the "self" IPs when checking the
peers list.

This behavior is reinforced in existing code in
[`internal/peer/redis.go`](https://github.com/honeycombio/refinery/blob/main/internal/peer/redis.go#L291-L298)
where `RedisIdentifier` is used to create a URL with the value being the
host.