Track span count and optionally add it to root (#532)
## Which problem is this PR solving?

Adds the ability to add the trace's span count to the root span of the
trace, and does it in a way that allows the rules engine to query it as
`meta.span_count`. This would allow rules that depend on span count --
for example, a rule that adjusts sample rate so that larger traces are
kept less often.

Note that this value is not available if the root span hasn't arrived
when the TraceTimeout completes; at this point, the rules will be
evaluated without a `span_count` because there is no root span to put it
on.

If the root span arrives late, it will be updated with the total number
of spans that have arrived at that point, so that the value stored in
Honeycomb is close to correct. If more spans arrive later, they will not
update the value in the root span.

Closes #524  
Closes #483