feat: Implement alternative sharding using rendezvous hash to improve dynamic scalability (#570)
## Which problem is this PR solving?

The legacy code in Refinery does a deterministic sharding calculation;
for every span that comes in, we hash the traceID with SHA1, truncate to
generate a 32-bit number, then divide the range by N if there are N
shards. If there are 4 shards, the lowest values go to shard 0, the next
block goes to shard 1, etc.

The problem is that if the number of shards changes from 4 to 5, 1/5 of
the shards that would have gone to 0 now go to 1. 2/5 of shard 1 now
goes to shard 2; 3/5 of shard 2 (more than half!) goes to shard 3, and
so forth. On average, half the shards move.

The practical result of this is that when a span arrives just after a
change in shard count, there’s a chance that it will end up on a
different shard from other spans in the same trace that arrived earlier.
The two shards might make different sampling decisions, and therefore we
could end up with missing spans in Honeycomb.

A technique called "[Rendezvous
Hashing](https://en.wikipedia.org/wiki/Rendezvous_hashing)" uses a
double hash — both the traceID and the shardID are hashed, and then we
do a calculation to find which trace should go with which shard.

There are several algorithms for doing this; all the algorithms seem to
be roughly similar in terms of their ability to balance the load among
shards. The randomness of traceIDs can lead to variation anyway, so it’s
mainly a probabalistic effect.

With one of these algorithms, adding or dropping a new shard usually
affects approximately 1/N of the existing traces, where N is the number
of shards. This is far better than half! Sometimes, random chance will
be unlucky, and it might be closer to 2/N, but in general it's still
quite a bit better than the legacy algorithm.

Implementing the new algorithm reduces the probability of being on a
different shard from 1/2 to 1/N, which is most meaningful for large
installations where N might be 15, 40, or even something like 120. This
seems worthwhile.

Note that the point of these algorithms is to allow shard assignment
decisions to be made independently by different shards yet generate the
same results without requiring any synchronization between shards.

It should be noted that while the legacy sharder is O(1) in its
operation, the hash sharder is O(N) where N is the number of partitions
(which is either 50 or the number of shards, whichever is greater). It
still runs in under a microsecond even with 128 shards.

## Short description of the changes

- Add configuration to choose legacy or hash sharding strategy (default
to legacy)
- Update example config to explain it
- Add new sharding strategy
- Add mechanism to select a sharding strategy based on the config
- Add tests to show that it works 
- Add benchmarks to show that it's performant enough